Replace custom lazy loading with lazy_loader, new dependency
