+++ src/kikuchipy/__init__.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-from typing import Union
+import lazy_loader
 
 # Initial committer first, then sorted by line contributions
 credits = [
@@ -31,7 +31,6 @@ credits = [
     "Magnus Nord",
 ]
 __version__ = "0.11.dev0"
-
 
 # Attempt (and fail) import of optional dependencies only once
 try:
@@ -74,77 +73,6 @@ except:  # pragma: no cover
     # LogicError, but we also want to catch import errors here
     _pyopencl_context_available = False
 
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
 
-def set_log_level(level: Union[int, str]):  # pragma: no cover
-    """Set level of kikuchipy logging messages.
-
-    Parameters
-    ----------
-    level
-        Any value accepted by :meth:`logging.Logger.setLevel()`. Levels
-        are ``"DEBUG"``, ``"INFO"``, ``"WARNING"``, ``"ERROR"`` and
-        ``"CRITICAL"``.
-
-    Notes
-    -----
-    See https://docs.python.org/3/howto/logging.html.
-
-    Examples
-    --------
-    Note that you might have to set the logging level of the root stream
-    handler to display kikuchipy's debug messages, as this handler might
-    have been initialized by another package
-
-    >>> import logging
-    >>> logging.root.handlers[0]  # doctest: +SKIP
-    <StreamHandler <stderr> (INFO)>
-    >>> logging.root.handlers[0].setLevel("DEBUG")
-
-    >>> import kikuchipy as kp
-    >>> kp.set_log_level("DEBUG")
-    >>> s = kp.data.nickel_ebsd_master_pattern_small()
-    >>> s.set_signal_type("EBSD")  # doctest: +SKIP
-    DEBUG:kikuchipy.signals._kikuchi_master_pattern:Delete custom attributes when setting signal type
-    """
-    import logging
-
-    logging.basicConfig()
-    logging.getLogger("kikuchipy").setLevel(level)
-
-
-__all__ = [
-    "__version__",
-    "_pyebsdindex_installed",
-    "_pyvista_installed",
-    "data",
-    "detectors",
-    "draw",
-    "filters",
-    "imaging",
-    "indexing",
-    "io",
-    "load",
-    "pattern",
-    "set_log_level",
-    "signals",
-    "simulations",
-]
-
-
-def __dir__():
-    return sorted(__all__)
-
-
-def __getattr__(name):
-    _import_mapping = {
-        "load": "io._io",
-    }
-    if name in __all__:
-        import importlib
-
-        if name in _import_mapping.keys():
-            import_path = f"{__name__}.{_import_mapping.get(name)}"
-            return getattr(importlib.import_module(import_path), name)
-        else:  # pragma: no cover
-            return importlib.import_module("." + name, __name__)
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+del lazy_loader
