+++ src/kikuchipy/__init__.pyi
@@ -15,16 +15,34 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-"""Tools for indexing of EBSD patterns by matching to a dictionary of
-simulated patterns.
+from . import (
+    data,
+    detectors,
+    draw,
+    filters,
+    imaging,
+    indexing,
+    io,
+    pattern,
+    signals,
+    simulations,
+)
+from .io._io import load
+from .logging import set_log_level
 
-Some of these tools are used in
-:meth:`~kikuchipy.signals.EBSD.dictionary_indexing`.
-"""
-
-import lazy_loader
-
-__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
-
-
-del lazy_loader
+__all__ = [
+    # Functions
+    "load",
+    "set_log_level",
+    # Modules
+    "data",
+    "detectors",
+    "draw",
+    "filters",
+    "imaging",
+    "indexing",
+    "io",
+    "pattern",
+    "signals",
+    "simulations",
+]
