+++ src/kikuchipy/_util/__init__.py
@@ -26,4 +26,8 @@ not for users.
     warning.
 """
 
-from kikuchipy._util._deprecated import deprecated, deprecated_argument
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
+
+del lazy_loader
