+++ src/kikuchipy/draw/__init__.py
@@ -17,27 +17,9 @@
 
 """Tools for use in plotting of signals."""
 
-__all__ = [
-    "get_rgb_navigator",
-    "plot_pattern_positions_in_map",
-]
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
 
 
-def __dir__():
-    return sorted(__all__)
-
-
-def __getattr__(name):
-    _import_mapping = {
-        "get_rgb_navigator": "_navigators",
-        "plot_pattern_positions_in_map": "_plot_pattern_positions_in_map",
-    }
-    if name in __all__:
-        import importlib
-
-        if name in _import_mapping.keys():
-            import_path = f"{__name__}.{_import_mapping.get(name)}"
-            return getattr(importlib.import_module(import_path), name)
-        else:  # pragma: no cover
-            return importlib.import_module("." + name, __name__)
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+del lazy_loader
