+++ src/kikuchipy/imaging/vbse.py
@@ -27,8 +27,9 @@ from hyperspy.roi import BaseInteractiveROI, RectangularROI
 import numpy as np
 
 from kikuchipy._util._transfer_axes import _transfer_navigation_axes_to_signal_axes
-from kikuchipy.pattern import rescale_intensity
-from kikuchipy.signals import EBSD, LazyEBSD, VirtualBSEImage
+from kikuchipy.pattern._pattern import rescale_intensity
+from kikuchipy.signals.ebsd import EBSD, LazyEBSD
+from kikuchipy.signals.virtual_bse_image import VirtualBSEImage
 
 
 class VirtualBSEImager:
