+++ src/kikuchipy/indexing/__init__.pyi
@@ -1,0 +1,42 @@
+# Copyright 2019-2024 The kikuchipy developers
+#
+# This file is part of kikuchipy.
+#
+# kikuchipy is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# kikuchipy is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
+
+from ._hough_indexing import xmap_from_hough_indexing_data
+from ._merge_crystal_maps import merge_crystal_maps
+from ._orientation_similarity_map import orientation_similarity_map
+from ._refinement._refinement import (
+    compute_refine_orientation_projection_center_results,
+    compute_refine_orientation_results,
+    compute_refine_projection_center_results,
+)
+from .similarity_metrics._normalized_cross_correlation import (
+    NormalizedCrossCorrelationMetric,
+)
+from .similarity_metrics._normalized_dot_product import NormalizedDotProductMetric
+from .similarity_metrics._similarity_metric import SimilarityMetric
+
+__all__ = [
+    "NormalizedCrossCorrelationMetric",
+    "NormalizedDotProductMetric",
+    "SimilarityMetric",
+    "compute_refine_orientation_projection_center_results",
+    "compute_refine_orientation_results",
+    "compute_refine_projection_center_results",
+    "merge_crystal_maps",
+    "orientation_similarity_map",
+    "xmap_from_hough_indexing_data",
+]
