+++ src/kikuchipy/indexing/similarity_metrics/__init__.py
@@ -14,16 +14,3 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
-
-# isort: off
-from ._similarity_metric import SimilarityMetric
-from ._normalized_cross_correlation import NormalizedCrossCorrelationMetric
-from ._normalized_dot_product import NormalizedDotProductMetric
-
-# isort: on
-
-__all__ = [
-    "NormalizedCrossCorrelationMetric",
-    "NormalizedDotProductMetric",
-    "SimilarityMetric",
-]
