+++ src/kikuchipy/io/plugins/__init__.pyi
@@ -15,16 +15,34 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-"""Tools for indexing of EBSD patterns by matching to a dictionary of
-simulated patterns.
+from . import (
+    bruker_h5ebsd,
+    ebsd_directory,
+    edax_binary,
+    edax_h5ebsd,
+    emsoft_ebsd,
+    emsoft_ebsd_master_pattern,
+    emsoft_ecp_master_pattern,
+    emsoft_tkd_master_pattern,
+    kikuchipy_h5ebsd,
+    nordif,
+    nordif_calibration_patterns,
+    oxford_binary,
+    oxford_h5ebsd,
+)
 
-Some of these tools are used in
-:meth:`~kikuchipy.signals.EBSD.dictionary_indexing`.
-"""
-
-import lazy_loader
-
-__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
-
-
-del lazy_loader
+__all__ = [
+    "bruker_h5ebsd",
+    "ebsd_directory",
+    "edax_binary",
+    "edax_h5ebsd",
+    "emsoft_ebsd",
+    "emsoft_ebsd_master_pattern",
+    "emsoft_ecp_master_pattern",
+    "emsoft_tkd_master_pattern",
+    "kikuchipy_h5ebsd",
+    "nordif",
+    "nordif_calibration_patterns",
+    "oxford_binary",
+    "oxford_h5ebsd",
+]
