+++ src/kikuchipy/io/plugins/edax_binary.py
@@ -27,7 +27,7 @@ import warnings
 import dask.array as da
 import numpy as np
 
-from kikuchipy.signals.util import get_chunking
+from kikuchipy.signals.util._dask import get_chunking
 
 __all__ = ["file_reader"]
