+++ src/kikuchipy/io/plugins/edax_h5ebsd.py
@@ -23,7 +23,7 @@ from typing import List, Union
 import h5py
 from orix.crystal_map import CrystalMap
 
-from kikuchipy.detectors import EBSDDetector
+from kikuchipy.detectors.ebsd_detector import EBSDDetector
 from kikuchipy.io.plugins._h5ebsd import H5EBSDReader, _hdf5group2dict
 
 __all__ = ["file_reader"]
