+++ src/kikuchipy/io/plugins/oxford_h5ebsd.py
@@ -26,7 +26,7 @@ import h5py
 import numpy as np
 from orix.crystal_map import CrystalMap
 
-from kikuchipy.detectors import EBSDDetector
+from kikuchipy.detectors.ebsd_detector import EBSDDetector
 from kikuchipy.io.plugins._h5ebsd import H5EBSDReader, _hdf5group2dict
 
 __all__ = ["file_reader"]
