+++ src/kikuchipy/pattern/__init__.py
@@ -17,44 +17,9 @@
 
 """Single pattern processing (used by signals)."""
 
-__all__ = [
-    "chunk",
-    "fft",
-    "fft_filter",
-    "fft_frequency_vectors",
-    "fft_spectrum",
-    "get_dynamic_background",
-    "get_image_quality",
-    "ifft",
-    "normalize_intensity",
-    "remove_dynamic_background",
-    "rescale_intensity",
-]
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
 
 
-def __dir__():
-    return sorted(__all__)
-
-
-def __getattr__(name):
-    _import_mapping = {
-        "fft": "_pattern",
-        "fft_filter": "_pattern",
-        "fft_frequency_vectors": "_pattern",
-        "fft_spectrum": "_pattern",
-        "get_dynamic_background": "_pattern",
-        "get_image_quality": "_pattern",
-        "ifft": "_pattern",
-        "normalize_intensity": "_pattern",
-        "remove_dynamic_background": "_pattern",
-        "rescale_intensity": "_pattern",
-    }
-    if name in __all__:
-        import importlib
-
-        if name in _import_mapping.keys():
-            import_path = f"{__name__}.{_import_mapping.get(name)}"
-            return getattr(importlib.import_module(import_path), name)
-        else:  # pragma: no cover
-            return importlib.import_module("." + name, __name__)
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+del lazy_loader
