+++ src/kikuchipy/signals/util/__init__.py
@@ -19,29 +19,9 @@
 :class:`~dask.array.Array` and other array tools.
 """
 
-__all__ = [
-    "get_chunking",
-    "get_dask_array",
-    "grid_indices",
-]
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
 
 
-def __dir__():
-    return sorted(__all__)
-
-
-def __getattr__(name):
-    _import_mapping = {
-        "get_chunking": "_dask",
-        "get_dask_array": "_dask",
-        "grid_indices": "array_tools",
-    }
-    if name in __all__:
-        import importlib
-
-        if name in _import_mapping.keys():
-            import_path = f"{__name__}.{_import_mapping.get(name)}"
-            return getattr(importlib.import_module(import_path), name)
-        else:  # pragma: no cover
-            return importlib.import_module("." + name, __name__)
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+del lazy_loader
