+++ src/kikuchipy/signals/util/__init__.pyi
@@ -15,16 +15,11 @@
 # You should have received a copy of the GNU General Public License
 # along with kikuchipy. If not, see <http://www.gnu.org/licenses/>.
 
-"""Tools for indexing of EBSD patterns by matching to a dictionary of
-simulated patterns.
+from ._dask import get_chunking, get_dask_array
+from .array_tools import grid_indices
 
-Some of these tools are used in
-:meth:`~kikuchipy.signals.EBSD.dictionary_indexing`.
-"""
-
-import lazy_loader
-
-__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
-
-
-del lazy_loader
+__all__ = [
+    "get_chunking",
+    "get_dask_array",
+    "grid_indices",
+]
