+++ src/kikuchipy/simulations/__init__.py
@@ -21,6 +21,10 @@ zone axes.
 
 import logging
 
+import lazy_loader
+
+__getattr__, __dir__, __all__ = lazy_loader.attach_stub(__name__, __file__)
+
 
 class DisableMatplotlibWarningFilter(logging.Filter):  # pragma: no cover
     # Filter to suppress warnings with the below warning message
@@ -38,28 +42,4 @@ class DisableMatplotlibWarningFilter(logging.Filter):  # pragma: no cover
 
 logging.getLogger("matplotlib.text").addFilter(DisableMatplotlibWarningFilter)
 
-
-__all__ = [
-    "GeometricalKikuchiPatternSimulation",
-    "KikuchiPatternSimulator",
-]
-
-
-def __dir__():
-    return sorted(__all__)
-
-
-def __getattr__(name):
-    _import_mapping = {
-        "GeometricalKikuchiPatternSimulation": "_kikuchi_pattern_simulation",
-        "KikuchiPatternSimulator": "kikuchi_pattern_simulator",
-    }
-    if name in __all__:
-        import importlib
-
-        if name in _import_mapping.keys():
-            import_path = f"{__name__}.{_import_mapping.get(name)}"
-            return getattr(importlib.import_module(import_path), name)
-        else:  # pragma: no cover
-            return importlib.import_module("." + name, __name__)
-    raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
+del lazy_loader, logging
